requirements.txt
aiohttp
argparse
requests
ruamel.yaml
